refactor(Typing): Improve typing
