+++ packages/contentful-apps/redirects/src/components/index.tsx
@@ -1,4 +1,3 @@
 export { default as Form } from './Form'
 export { default as Modal } from './Modal'
-export type { FormErrors, Redirect, Status } from './primitives'
 export { default as Table } from './Table'
