+++ packages/contentful-apps/redirects/src/types/Redirect.ts
@@ -1,6 +1,6 @@
 export type Redirect = {
   from: string
   to: string
-  status: string | number
+  status: number
   date?: number
 }
