+++ packages/contentful-apps/redirects/src/types/index.tsx
